+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Preserve `Eloquent\Builder` when overriding the default ordering of dataTables when using `EloquentEngine`.
 - Preserve `Eloquent\Builder` when using filterColumn api. Allows us to use model scope and any eloquent magics.
 - Fractal integration extracted to own plugin [laravel-datatables-fractal](https://github.com/yajra/laravel-datatables-fractal).
+- Raw output are always passed on transformer instance.
 
 ### REMOVED
 - Remove `filterColumn` api magic query method in favor of closure.